# packages/curatarr/docs/site/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Curatarr API
  version: 0.2.0
  description: |
    Curatarr HTTP API reference.

    Notes:
    - This spec reflects endpoints currently mounted in `src/server/app.ts`.
    - Several response payloads are intentionally flexible (`additionalProperties: true`) while the project is in active MVP iteration.
servers:
  - url: https://curatarr.video
    description: Production
  - url: http://localhost:7474
    description: Local
...
paths:
...
  /api/verify/start:
    post:
      tags: [Verify]
      summary: Start deep check queue
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                concurrency: { type: integer, minimum: 1, maximum: 8 }
                fileIds:
                  type: array
                  items: { type: integer }
                rescan: { type: boolean }
                budgetSeconds: { type: integer, minimum: 30, maximum: 3600 }
      responses:
        '200':
          description: Verify started
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnyObject' }
        '409':
          description: Verify already running
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /api/verify/failures:
    get:
      tags: [Verify]
      summary: List verify failures
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Verify failures
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnyObject' }
...
components:
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
  schemas:
    AnyObject:
      type: object
      additionalProperties: true
    ErrorResponse:
      type: object
      properties:
        error:
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
        detail:
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
      additionalProperties: true
...